split into partial class instead of ifdefs
