+++ src/BoxKite.Twitter/BoxKite.Twitter.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>netstandard1.4;net452;Xamarin.iOS10;Xamarin.Mac20;MonoAndroid70</TargetFrameworks>
+    <TargetFrameworks>netstandard1.4;net452;Xamarin.iOS10;Xamarin.Mac20;MonoAndroid70;uap10.0</TargetFrameworks>
     <AssemblyName>BoxKite.Twitter</AssemblyName>
     <RootNamespace>BoxKite.Twitter</RootNamespace>
     <authors>Nick Hodge</authors>
@@ -15,23 +15,36 @@
   </PropertyGroup>
 
  <ItemGroup>
+    <!--see https://github.com/onovotny/MSBuildSdkExtras  -->
+    <PackageReference Include="MSBuild.Sdk.Extras" Version="1.0.5" PrivateAssets="all" />
+</ItemGroup>
+
+ <ItemGroup>
     <Compile Remove="Platforms\**\*.cs" />
     <None Include="Platforms\**\*.cs" />
     <PackageReference Include="System.Reactive" Version="3.1.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />    
+    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
+
+    <!-- System.Runtime.Serialization.Primitives reference not needed when netstandard 2.0 comes out https://stackoverflow.com/a/43956572 -->
+    <PackageReference Include="System.Runtime.Serialization.Primitives" Version="4.3.0" />
 </ItemGroup>
 
-<ItemGroup Condition="'$(TargetFramework)'!='netstandard1.4'">
-  <Compile Include="Platforms\shared\**\*.cs" />    
+<ItemGroup>
+  <Compile Include="Platforms\shared\**\*.cs" />
 </ItemGroup>
 
- <ItemGroup Condition=" '$(TargetFramework)' == 'net45' ">
+<ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.4'">
+  <Compile Include="Platforms\netstandard1.4\**\*.cs" />
+</ItemGroup>
+
+ <ItemGroup Condition=" '$(TargetFramework)' == 'net452' ">
     <Compile Include="Platforms\net452\**\*.cs" />
   </ItemGroup>
 
-  <!--<ItemGroup Condition=" '$(TargetFramework)' == 'uap10.0' ">
-    <Compile Include="Platforms\uap10.0\**\*.cs" />
-  </ItemGroup>-->
+  <ItemGroup Condition=" '$(TargetFramework)' == 'uap10.0' ">
+    <Compile Include="Platforms\uap\**\*.cs" />
+    <PackageReference Include="Microsoft.NETCore.UniversalWindowsPlatform" Version="5.3.3" /> 
+  </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'Xamarin.iOS10' ">
     <Compile Include="Platforms\ios\**\*.cs" />
@@ -44,4 +57,8 @@
   <ItemGroup Condition=" '$(TargetFramework)' == 'MonoAndroid70' ">
     <Compile Include="Platforms\android\**\*.cs" />
   </ItemGroup>  
+
+  <!--see https://github.com/onovotny/MSBuildSdkExtras  -->
+  <Import Project="$(MSBuildSDKExtrasTargets)" Condition="Exists('$(MSBuildSDKExtrasTargets)')" />
 </Project>
+
