+++ src/BoxKite.Twitter/Platforms/shared/Connection/TwitterConnection.cs
@@ -41,140 +41,6 @@ namespace BoxKite.Twitter
         public IUserStream UserStream { get; set; }
         public ISearchStream SearchStream { get; set; }
 
-#if(PORTABLE)
-        public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, IPlatformAdaptor platformAdaptor = null)
-        {
-            PlatformAdaptor = platformAdaptor;
-            ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-            UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-        }
-
-        public TwitterConnection(TwitterCredentials twitterCredentials, IPlatformAdaptor platformAdaptor = null)
-        {
-            PlatformAdaptor = platformAdaptor;
-            TwitterCredentials = twitterCredentials;
-            UserSession = BuildUserSession();
-        }
-
-        // ReSharper disable once UnusedParameter.Local
-        public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, string xauthusername, string xauthpassword, IPlatformAdaptor platformAdaptor = null)
-        {
-            PlatformAdaptor = platformAdaptor;
-            UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-        }
-#endif
-#if (WINDOWSDESKTOP)
-        public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret)
-        {
-            PlatformAdaptor = new DesktopPlatformAdaptor();
-            ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-            UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-         }
-
-       public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, IPlatformAdaptor platformAdaptor = null)
-        {
-            PlatformAdaptor = platformAdaptor;
-            ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-            UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-        }
-
-        public TwitterConnection(TwitterCredentials twitterCredentials)
-        {
-            PlatformAdaptor = new DesktopPlatformAdaptor();
-            TwitterCredentials = twitterCredentials;
-            UserSession = BuildUserSession();
-        }
-
-        public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, string xauthusername, string xauthpassword)
-        {
-            PlatformAdaptor = new DesktopPlatformAdaptor();
-            UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-        }
-#endif
-#if(WINDOWS_UWP)
-        public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret)
-        {
-            PlatformAdaptor = new WinUWPPlatformAdaptor();
-            ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-            UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-         }
-
-       public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, IPlatformAdaptor platformAdaptor = null)
-        {
-            PlatformAdaptor = platformAdaptor;
-            ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-            UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-        }
-
-        public TwitterConnection(TwitterCredentials twitterCredentials)
-        {
-            PlatformAdaptor = new WinUWPPlatformAdaptor();
-            TwitterCredentials = twitterCredentials;
-            UserSession = BuildUserSession();
-        }
-
-        public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, string xauthusername, string xauthpassword)
-        {
-            PlatformAdaptor = new WinUWPPlatformAdaptor();
-            UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-         } 
-#endif
-#if(IOS)
-		public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret)
-		{
-			PlatformAdaptor = new IOSPlatformAdaptor();
-			ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-			UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-		}
-
-		public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, IPlatformAdaptor platformAdaptor = null)
-		{
-		PlatformAdaptor = platformAdaptor;
-		ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-		UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-		}
-
-		public TwitterConnection(TwitterCredentials twitterCredentials)
-		{
-			PlatformAdaptor = new IOSPlatformAdaptor();
-			TwitterCredentials = twitterCredentials;
-			UserSession = BuildUserSession();
-		}
-
-		public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, string xauthusername, string xauthpassword)
-		{
-			PlatformAdaptor = new IOSPlatformAdaptor();
-			UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-		} 
-#endif
-#if(ANDROID)
-		public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret)
-		{
-			PlatformAdaptor = new AndroidPlatformAdaptor();
-			ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-			UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-		}
-
-		public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, IPlatformAdaptor platformAdaptor = null)
-		{
-		    PlatformAdaptor = platformAdaptor;
-		    ApplicationSession = BuildApplicationSession(twitterConsumerKey, twitterConsumerSecret);
-		    UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-		}
-
-		public TwitterConnection(TwitterCredentials twitterCredentials)
-		{
-			PlatformAdaptor = new AndroidPlatformAdaptor();
-			TwitterCredentials = twitterCredentials;
-			UserSession = BuildUserSession();
-		}
-
-		public TwitterConnection(string twitterConsumerKey, string twitterConsumerSecret, string xauthusername, string xauthpassword)
-		{
-			PlatformAdaptor = new AndroidPlatformAdaptor();
-			UserSession = BuildUserSession(twitterConsumerKey, twitterConsumerSecret);
-		} 
-#endif
         private IApplicationSession BuildApplicationSession(string twitterConsumerKey, string twitterConsumerSecret)
         {
             return ApplicationSession ?? new ApplicationSession(twitterConsumerKey, twitterConsumerSecret);
