+++ src/BoxKite.Twitter/Platforms/uap/WinUWPPlatformAdaptor.cs
